pyproject.toml
[tool.poetry]
name = "image-copier"
version = "0.1.0"
description = ""
authors = ["Jannik Bach <jannik.bach@inovex.de>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.2.0"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "*"
ruff = "*"
ipython = "*"
pyinstaller = "^6.3.0"

[tool.poetry.scripts]
build = "src.pyinstaller:install"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E", # pycodestyle error
    "W", # pycodestyle warning
    "F", # pyflakes
    "A", # flakes8-builtins
    "COM", # flakes8-commas
    "C4", # flake8-comprehensions
    "Q", # flake8-quotes
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "I", # isort
    "N", # pep8 naming
    "UP", # pyupgrade  
    "S", # bandit
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in select subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
# 5. Use single quotes for non-triple-quoted strings.
quote-style = "single"
